requirements.txt
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
google-cloud-pubsub==2.6.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
httplib2==0.19.1
libcst==0.3.19
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests
rsa==4.7.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
uritemplate==3.0.1